"Fork" the CSS for vue-tel-input to fix the postcss incompatibilities
The post-css loader built into nuxt was throwing a hissy about how the
vue-tel-input devs built the CSS - specifically it does not like
custom props declared anywhere but :root and
if it can't find it at :root, it doesn't know what to do about it.

Since the vue-tel-input CSS did not include fallback values for the
var() function, it was throwing warnings every time the dev server
reloaded. I fixed these issues.