Re-enable Vuex-Persist and clean up error handling
Turned Vuex-persist back on to allow for local storage of session
information. This will allow the page to maintain
the data in case the user accidentally reloads
or navigates away.

I also cleaned up a lot of the error handling stuff sprinkled
throughout the pages. I think the result is a more elegant,
 easier to follow implementation. Not perfect, but better.