Add handler for user re-submitting verification for an in-progress email.
If the user submits their email a second time after a successful
submission, Salesforce will reissue the email verification
email.

The email.vue page will get a response of 200 (instead of 204)
and redirect to the email-thankyou page with a query param 'cd'

I hard coded this "code", which will prompt the user that a verification
was resent.