Add RESET_STATE and adjust email.vue query string check
Added a RESET_STATE mutator to the Vuex Store. I can't simply remove
the localStorage item by key because the plugin simply
re-adds it from memory. So Instead I'm clearing the state to
a default.

I call the RESET_STATE mutator from the email page (if you're
verifying the email address,  you're probably starting from scratch)
and from the Summary page completion success.

I also adjusted the query string check on email.vue. I moved
the check to Mounted instead of Fetch and I put in a check
to handle the case of an actual email address being used
as the query param instead of a base64 encoded string.