Update error handling for initial page load
I wanted more granular error handling on the Contact page,
which should be initiating the GUID verification step when the user
first comes to the page from the link mailed to them.

Additionally, I discovered that I was stupidly saving the error state
in the persisted data store, which means the page was not able
to correctly respond to changing states that came from server-side.

I created a filter function for vuex-persisteddata to exclude the error
state. This seems to have solved the issue.